+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+-   Migrate to Typescript
 -   Breaking: Update [globby](https://github.com/sindresorhus/globby)
 
 ## [3.0.0] - 2019-01-07